hr-platform/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.21
openai==1.2.0
pymupdf==1.22.5
python-docx==0.8.11
python-multipart==0.0.6
aiofiles==23.2.1
